// frontend/src/services/api/models/EmployeeSalaryCreate.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type EmployeeSalaryCreate = {
    employee_id: number;
    component_id: number;
    amount: number;
    effective_date: string;
    end_date?: (string | null);
};