// frontend/src/services/api/models/EmployeeSalaryUpdate.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type EmployeeSalaryUpdate = {
    employee_id?: (number | null);
    component_id?: (number | null);
    amount?: (number | null);
    effective_date?: (string | null);
    end_date?: (string | null);
};