// frontend/src/services/api/models/EmployeeSalaryInDB.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type EmployeeSalaryInDB = {
    employee_id: number;
    component_id: number;
    amount: number;
    effective_date: string;
    end_date?: (string | null);
    id: number;
    created_at: string;
    updated_at: string;
};